+++ cpclib-bndbuild/src/runners/mod.rs
@@ -15,6 +15,7 @@ pub mod emulator;
 pub mod fade;
 pub mod fs;
 pub mod hideur;
+pub mod hxcfe;
 pub mod img2cpc;
 pub mod locomotive;
 pub mod snapshot;